4.4. Максимум из трех чисел[#131521]

+++ chapter_001/src/main/java/ru/job4j/condition/Max.java
@@ -1,7 +1,7 @@
 package ru.job4j.condition;
 
 public class Max {
-    public int max(int left, int right) {
+    public double max(double left, double right) {
         return left >= right ? left : right;
     }
 }
